20060313: add code to convert k37 to k23

+++ tcb_ploan/prog/loan_app.h
@@ -214,7 +214,7 @@ class LoanApp {
 //    int  write_result_others(int decision);
 //    int  write_result_declined(int decision);
 //    int  write_result_approved(int decision);
-    int  write_result(int decision, double npv_value, int unsecured_amt);
+    int  write_result(int decision, int npv_value, int unsecured_amt);
     double calculate_npv(double delta_apr);
     double find_lowest_rate (double offset, double delta_r);
     void set_lowest_rate (double delta_rate);
